Seperate some input system logic into seperate files Add default implementation for InputSystem hhoks: `func/input_setup`
