+++ inc/input.h
@@ -6,7 +6,7 @@
 /*   By: clovell <clovell@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/25 17:18:00 by clovell           #+#    #+#             */
-/*   Updated: 2023/11/22 21:39:26 by clovell          ###   ########.fr       */
+/*   Updated: 2023/11/23 01:08:35 by clovell          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef INPUT_H
@@ -22,7 +22,7 @@ typedef struct s_inputctx t_inputctx;
 
 struct s_inputctx
 {
-	uint8_t key_state[256];
+	uint8_t key_state[KEY_REG_MAX];
 	uint8_t special_state[256];
 	uint8_t mouse_state[5];
 
@@ -39,5 +39,13 @@ bool	input_keyup(t_game *game, int key);
 bool	input_keyheld(t_game *game, int key);
 void	input_process(t_game *game);
 uint8_t	*input_get_state(t_game *game, int *key);
+
+
+void	input_setup(t_game *game);
+void	input_cb_key_press(int key, t_game *game);
+void	input_cb_key_release(int key, t_game *game);
+void	input_cb_mouse_press(int button, int x, int y, t_game *game);
+void	input_cb_mouse_release(int button, int x, int y, t_game *game);
+void	input_cb_mouse_move(int x, int y, t_game *game);
 
 #endif
